refactor: move the playlist selection modal to its own component
